Moved files location to keep all together
