+++ src/main/scala/beam/utils/csv/NetworkXmlToCSV.scala
@@ -1,4 +1,4 @@
-package beam.utils
+package beam.utils.csv
 
 import java.io.{BufferedReader, File, PrintWriter}
