+++ src/main/scala/beam/utils/csv/PopulationXml2CsvConverter.scala
@@ -1,4 +1,4 @@
-package scripts
+package beam.utils.csv
 
 import java.io.File
