+++ src/test/scala/beam/utils/NetworkXmlToCsvSpec.scala
@@ -1,5 +1,6 @@
 package beam.utils
 
+import beam.utils.csv.NetworkXmlToCSV
 import org.scalatest.WordSpecLike
 
 class NetworkXmlToCsvSpec extends WordSpecLike {
